feature #507: Made sunstone.js unaware of the existence of dataTables.
DataTables are a very particular object which is part of the
plugins. Therefore instead of having sunstone.js take a dataTable and find out
which elements are checked in order to run an action on them, this is now work
to be done by the plugins. They can use a helper called
getSelectedNodes(dataTable) to do that. Actions now define an "elements"
function which returns an array of elements, instead of having a "dataTable"
returning the dataTable of the plugin. runActionOnDatatableNodes not needed anymore.