Feature #799: Refactor ozones.js code
This commit refactors the code in ozone.js.

In order to allow as much refactoring as possible, server and zones plugin have been modified so that a better consistency is achieved:

 - "vnet" replaces "vn" to identify virtual networks resource.
 - "vmtemplate" replaces "template" to identify virtual machine templates.

In the server side, all "vnet", "vn" / "vmtemplate", "template" produce the same result, so that the CLI spec and oZones doc are not affected by this change.

Note: Additional changes on OzonesServer.rb correspond to whitespace cleanup.