Feature #1004: Add internationalization support for Sunstone
This commit enables internationalization support for the Sunstone interface.

Language is chosen according to the variable LANG in user's template. This variable is automaticly set when users chooses a new language from the select field on top, so that it is preserved on future logins.

If no LANG is set, language is chosen according to the :lang: variable of the configuration. Otherwise it defaults to English.

Special thanks to ZAO VIVOSS i OI for contributing the initial patch for OpenNebula Sunstone 3.0 and the first translation: russian. The contents from this patch have been partially applied with the necessary corrections and improvements to the master branch, which was ahead on many aspects to the 3.0 release.