Feature #1076: Add VNC support in SelfService.
When enabled in the occi server configuration file, UI users will be able to click the VNC icon that appears in the VM information.

Then the websockets proxy will be set up, provided that the machine has been configured with the appropiate GRAPHICS section etc. This must be done in the OCCI templates, and cannot be done by the UI user.

Wss sessions can be configured in the occi server configuration file. Unlike Sunstone, here they are transparent to the user and whenever they are enabled VNC sessions will be launched using wss:// automaticly. As such, it is not up to the user to choose the type of connection, and it fully depends on the server configuration.

Additionally the install_novnc.sh script has been updated and improved. The install.sh has been updated too.
(cherry picked from commit 4ccaf9704a5d415e4979bbf53dce263d2c1e7cdb)