feature #2289: monitor all VMs in EC2 IM
Get information from all EC2 instances in IM driver and provide
this information to OpenNebula so a VM poll is not executed per VM.
The monitoring will be much faster as VM is very expensive and the time
it takes to get info about one VM is very similar to getting information
about all VMs.

To keep track of the OpenNebula ID in EC2 machines a new tag is added
to EC2 instances (ONE_ID). This is used when monitoring VMs. VMs without
this tag wont be correctly reported to OpenNebula and will be monitored
with the old method (VM poll).

This version is kind of a PoC. The driver should be refactored so it
is similar to other drivers, that is, one script per action. The library
that holds the functionality will be held in a library file in VMM
driver as VMware drivers. Special care should be taken read
EC2 parameters as now is needed by both VMM and IM.