Fix 'one start' with existent DB but nonexistent auth file
The start script assumes that an existing database implies
credentials located in either one of these places:
  * $HOME/.one/one_auth or
  * $ONE_AUTH

Now, if the database was setup once correctly but the
credentials file does not exist or ONE_AUTH is not set
anymore, oned runs into a segmentation fault.

The change prevents this situation by always checking for
user credentials file wheter the DB exists or not.