bug #2723: fix use of --user with --cluster in the cli
--cluster param needs to create a client to search for the cluster. This
happens while the parameters are being parsed so it fails if there is no
ONE_AUTH. To make it work the parameters user, password and endpoint
are saved as OneHelper class variables as soon as they are parsed. This way
they can be used to create the XMLRPC client before all parameters are parsed.