bug #3315: use semantic versioning for remotes
onehost sync was using alphabetical comparison but this method does not work for this
comparison:

4.10.0 > 4.8.0

Now the versions are parsed by Gem::Version, that comes with rubygems and should
be installed in the frontend.

When the version can not be parsed it falls back to text comparison or forces update
if the comparison can not be made (one of the two versions is malformed).

More information about semantic versioning: http://semver.org/