Check if an given image is readable in oneimage.
When running "oneimage create --path .." the tool should check not only
for the existence but also for the readability of the given image.
If the image is unreadable (one user create the image an other tries to
upload it) the image will be created in ERROR state in the datastore.