Performance improvements for ssh TM_MAD driver
Accessing and transferring local files via ssh connection causes
significant unnecessary system load (CPU utilization). This patch
takes into consideration data locality and reduces system load by
up to 50 % when transferring large files between the local machine
and a remote machine.

Patch outline
=============
The local machine is identified by comparing the host part of source
and destination string with the local hostname. In case no match
is detected, old driver behavior is used. If a match is detected
(in source or destination string), the `$HOST:` prefix is removed
and `$PATH` is considered to be locally available. Where ssh-to-ssh
pipeline is used, one part of this pipeline will be executed locally
(not as a remote command through an ssh connection).

Modified files
==============
 src/tm_mad/ssh/clone
 src/tm_mad/ssh/cpds
 src/tm_mad/ssh/mv
 src/tm_mad/ssh/mvds