feature-3060: enable disk discard option for qemu-kvm
This is proposed solution for feature in Backlog #3060

By default this feature is not touching the xml configuration for backward compatibility.

There is global VM config in vmm/remotes/kvm/kvmrc named 'DEFAULT_ATTACH_DISCARD' that if set
will enforce the discard option to all VMs if there is no configuranito per template.

It is possible to enable the discard feature per VM in their template by adding
DISCARD=unmap
to enable and
DISCARD=ignore
to disable the discard