Feature #4215: Reverse NETWORK and NETWORK_ID priority
When requesting a vnet, NETWORK had priority over _ID.
If the NIC was taken from the VR, the NETWORK was set
but the _UNAME or _UID was not. In this case, VR
instantiate only worked for the owner of the VNet.

The priority of IMAGE and IMAGE_ID for disks is also
reversed, for consistency