tm/ceph {pre,post}migrate do nothing if not SYSTEM DS
The proposed solution partially addresses bug #4924

IMO in case of SYSTEM DS on shared filesystem the TM_MAD must be 'shared'.
In this case the tm/ceph/mv will not be called but then {pre,post}migrate
is called by migrate-other. In that case there is no work for Ceph.