B #5041: Avoid gathering KVM I/O statistics for saving VMs (#160)
Note: Branch 5.12 didn't really suffer from the similar problem in 5.10,
as here "saving" reason is considered as UNKNOWN state for which we
don't collect the I/O statistics. This change only aligns with previous
behaviour to consider "saving" as RUNNING state, and explicitly lists
few RUNNING reasons (including new "saving") to skip monitoring for.

It's more a polishment, than a real bugfix.