B #4047: allow resize for disks with snapshots"
This patch introduces a TM_MAD_CONF attribute
ALLOW_DISK_RESIZE_WITH_SNAPSHOT which specifies whether the driver
supports resizing of disks if they have snapshots. The attribute value
should be "YES" or "yes" to allow resize operation. Default is to deny
operation if the attribute is not present or is set as "NO" or "no".

+refs: #1260

Signed-off-by: Chirag Anand <anand.chirag@gmail.com>