M #-: nodeport post needs deactivate (#1237)
When resuming a VM powered-off from guest OS (and it is monitored as
powered-off by the OpenNebula monitor), it fails because 'ip route add'
fails when an already existing route is configured. If 'ip route add' is
changed by 'ip route replace' (idempotent) the problem is solved;
however, duplicated iptables rules are added. Therefore, to have this
working for all scenarios a deactivate is called on driver to clean the
VM route and the all the iptables rules previos to the call to activate.

Signed-off-by: Ricardo Diaz <rdiaz@opennebula.io>
(cherry picked from commit cc2ae9ef7bb54b30a9480dd4f56cd7d59286af03)