B #5513: Only match one LDAP match_user_regex entry (#5514)
When a match is made, the username is changed. It's then
inappropriate to attempt further matches because we're not using
the original username.

If multiple matches were made with the original username, there
could be no single modified username to return. An alternative
would be to return pairs of server and modified username.

Signed-off-by: Simon Arlott
(cherry picked from commit 721aa612326262bd1cb370743218b2632a67b306)