M #-: Fix iptables commands on versions < 1.6.1 and > 1.8.7
From version 1.8 iptbles front-end uses the nf_tables API to talk to the
kernel. One of the benefits is that is no longer needed to "wait", from
the man page:

"Because the xtables-nft tools use the nf_tables kernel API, rule
additions and deletions are always atomic. Unlike iptables-legacy,
iptables-nft -A .. will NOT need to retrieve the current ruleset from
the kernel, change it, and re-load the altered ruleset. Instead,
iptables-nft will tell the kernel to add one rule. For this reason, the
iptables-legacy --wait option is a no-op in iptables-nft."

This commit do no generate the --wait option from versions > 1.8.7

Signed-off-by: Kristian Feldsam <feldsam@gmail.com>
(cherry picked from commit 91ca9841a96433ad09bd833e2835067fa4a2d342)