F #6333: Fix path for multiple disks (1/2)
This is a partial fix for  backup snapshots for VMs with multiple disks.

It still fails for disks attached after the VM is created. In this case
the SOURCE of the disk is not replace by the "readlink", causing
problems to libvirt while resolving disks in the backingStore.

(cherry picked from commit 248bc50e16f6713e05e1687f13cfb76c11c7f57c)