F #6313: Add configurable backup path
This path refers to a location in the hypervisor FS, to temporally store backup
disks. The path is configured with the BACKUP_BASE_PATH variable in
datastore.conf file.

This commit also updates the references in the cancel action for the
rsync datastore

(cherry picked from commit 93ab45aec786435f0c0622638c43ed874f5a3f60)