B #6704: Perform SG and spoofing on input traffic
When the VM generates traffic addressed to the hypervisor it is handled
as routed (not bridged) and handle through the INPUT chain (not the
FORWARD chain). In this case, none of the iptables rules are applied to it so
bypassing all SG rules, including anti-spoofing ones.

This commit applies the egress rules to the VM traffic in this case.

L #-: Liniting of fw driver